requirements/dev.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements.txt --no-annotate --no-editable --no-hashes --only-group dev -o requirements/dev.txt
bandit==1.9.4
colorama==0.4.6 ; sys_platform == 'win32'
coverage==7.13.5
exceptiongroup==1.3.1 ; python_full_version < '3.11'
iniconfig==2.3.0
librt==0.8.1 ; platform_python_implementation != 'PyPy'
markdown-it-py==4.0.0
mdurl==0.1.2
mypy==1.19.1
mypy-extensions==1.1.0
packaging==26.0
pathspec==1.0.4
pluggy==1.6.0
psycopg==3.3.3
psycopg-binary==3.3.3 ; implementation_name != 'pypy'
pygments==2.20.0
pytest==9.0.2
pytest-qgis @ git+https://github.com/3liz/pytest-qgis.git@1f18d2d183f1b61888304c3e1e28684355288919
pyyaml==6.0.3
rich==14.3.3
ruff==0.15.8
semver==3.0.4
stevedore==5.7.0
tomli==2.4.1 ; python_full_version <= '3.11'
types-requests==2.33.0.20260327
typing-extensions==4.15.0
tzdata==2025.3 ; sys_platform == 'win32'
urllib3==2.6.3
vulture==2.16
